+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 
 - When the ping callable is called it must return a `Innmind\FileWatch\Continuation`
 
+### Removed
+
+- `Innmind\FileWatch\Stop`
+- `Innmind\FileWatch\Failed`
+
 ## 3.2.0 - 2023-11-12
 
 ### Added